refactor: define memory in main.rs
This is for compatibility with JIT compiler. JIT compiler would
return some memory map which can be executed, and it means the
local stack for the function which performs compilation is cleared.

So if I define the memory in the function, it would be cleared
after returning the memory map to be executed.

Putting the memory in higher level (in our case, in main function)
solves this problem directly.